Fix notification dispatch lint split
